+++ src/prefabs/datePickerInput.tsx
@@ -3,6 +3,7 @@ import {
   BeforeCreateArgs,
   Icon,
   prefab as makePrefab,
+  PrefabComponentOption,
 } from '@betty-blocks/component-sdk';
 import { DateTimePicker } from './structures/DateTimePicker';
 
@@ -223,7 +224,7 @@ const beforeCreate = ({
             configuration: {
               condition: {
                 type: 'SHOW',
-                option: 'actionProperty',
+                option: 'property',
                 comparator: 'EQ',
                 value: '',
               },
@@ -241,29 +242,31 @@ const beforeCreate = ({
           }));
 
           if (propertyBased) {
-            setOption(newPrefab.structure[0], 'actionProperty', (option) => ({
-              ...option,
-              value: {
-                modelProperty: propertyPath,
-                actionVariableId: result.variable.variableId,
-              },
-              configuration: {
-                condition: {
-                  type: 'HIDE',
-                  option: 'actionProperty',
-                  comparator: 'EQ',
-                  value: '',
+            setOption(
+              newPrefab.structure[0],
+              'property',
+              (originalOption: PrefabComponentOption) => ({
+                ...originalOption,
+                value: {
+                  id:
+                    result.isRelational && !result.isMultiRelational
+                      ? [propertyId, modelProperty.id]
+                      : propertyId,
+                  type: 'PROPERTY',
+                  name:
+                    result.isRelational && !result.isMultiRelational
+                      ? `{{ ${model?.name}.${name}.id }}`
+                      : `{{ ${model?.name}.${name} }}`,
                 },
-              },
-            }));
+              }),
+            );
           }
           if (validate()) {
             if (
               (selectedPrefab?.name === BettyPrefabs.UPDATE_FORM ||
-                ((selectedPrefab?.name === BettyPrefabs.CREATE_FORM ||
-                  selectedPrefab?.name === BettyPrefabs.FORM ||
-                  selectedPrefab?.name === BettyPrefabs.LOGIN_FORM) &&
-                  originalPrefab.name === BettyPrefabs.HIDDEN)) &&
+                selectedPrefab?.name === BettyPrefabs.CREATE_FORM ||
+                selectedPrefab?.name === BettyPrefabs.FORM ||
+                selectedPrefab?.name === BettyPrefabs.LOGIN_FORM) &&
               propertyId
             ) {
               const valueOptions = [