+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [2.125.0](https://github.com/bettyblocks/material-ui-component-set/compare/v2.124.3...v2.125.0) (2023-04-07)
+
+
+### Features
+
+* convert actionProperty to property option ([42bef12](https://github.com/bettyblocks/material-ui-component-set/commit/42bef12a61e287203acd94f294744317f9b114be))
+
+## [2.124.3](https://github.com/bettyblocks/material-ui-component-set/compare/v2.124.2...v2.124.3) (2023-03-31)
+
+
+### Bug Fixes
+
+* empty lable show uuid ([e2b8663](https://github.com/bettyblocks/material-ui-component-set/commit/e2b8663cd5c107ad408118195d5d8c8aa08382d6))
+
+## [2.124.2](https://github.com/bettyblocks/material-ui-component-set/compare/v2.124.1...v2.124.2) (2023-03-31)
+
+
+### Bug Fixes
+
+* decimal label value breaks autocomplete ([db5538e](https://github.com/bettyblocks/material-ui-component-set/commit/db5538e6e80c4252d74c9a4fd9bf6cdb8919a500))
+
+## [2.124.1](https://github.com/bettyblocks/material-ui-component-set/compare/v2.124.0...v2.124.1) (2023-03-31)
+
+
+### Bug Fixes
+
+* add enable/disable interaction to datepickers ([9630bd0](https://github.com/bettyblocks/material-ui-component-set/commit/9630bd0478e4c76a5aa77090b733c75ee0a56316))
+
 # [2.124.0](https://github.com/bettyblocks/material-ui-component-set/compare/v2.123.0...v2.124.0) (2023-03-28)